ai_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langgraph==0.1.19
langchain==0.2.6
langchain-google-genai==1.0.6
google-generativeai==0.5.4
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
